# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Dental Clinic Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /treatments:
    get:
      operationId: listTreatments
      tags: [treatments]
      summary: List treatments
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Treatment"
    post:
      operationId: createTreatment
      tags: [treatments]
      summary: Create treatment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTreatmentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Treatment"
  /treatments/{id}:
    put:
      operationId: updateTreatment
      tags: [treatments]
      summary: Update treatment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTreatmentRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Treatment"
    delete:
      operationId: deleteTreatment
      tags: [treatments]
      summary: Delete treatment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List staff
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Staff"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Create staff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
  /staff/{id}:
    put:
      operationId: updateStaff
      tags: [staff]
      summary: Update staff
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
    delete:
      operationId: deleteStaff
      tags: [staff]
      summary: Delete staff
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /reports/daily-revenue:
    get:
      operationId: getDailyRevenue
      tags: [reports]
      summary: Daily revenue report
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueReport"
  /reports/monthly-revenue:
    get:
      operationId: getMonthlyRevenue
      tags: [reports]
      summary: Monthly revenue report
      parameters:
        - name: month
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueReport"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Treatment:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        cost:
          type: number
        duration:
          type: integer
        description:
          type: string
      required:
        - id
        - name
        - cost
    CreateTreatmentRequest:
      type: object
      properties:
        name:
          type: string
        cost:
          type: number
        duration:
          type: integer
        description:
          type: string
      required:
        - name
        - cost
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        patientId:
          type: integer
        patientName:
          type: string
        date:
          type: string
        subtotal:
          type: number
        discount:
          type: number
        gst:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        createdAt:
          type: string
      required:
        - id
        - invoiceNumber
        - patientId
        - date
        - subtotal
        - total
        - paymentStatus
...
    Staff:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        phone:
          type: string
        email:
          type: string
        specialization:
          type: string
        joiningDate:
          type: string
        salary:
          type: number
        status:
          type: string
      required:
        - id
        - name
        - role
    CreateStaffRequest:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        phone:
          type: string
        email:
          type: string
        specialization:
          type: string
        joiningDate:
          type: string
        salary:
          type: number
        status:
          type: string
      required:
        - name
        - role
...
    RevenueReport:
      type: object
      properties:
        totalRevenue:
          type: number
        totalInvoices:
          type: integer
        paidAmount:
          type: number
        pendingAmount:
          type: number
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
      required:
        - totalRevenue
        - totalInvoices
        - paidAmount
        - pendingAmount
...