# lib/api-zod/src/generated/types/createTreatmentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dental Clinic Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTreatmentRequest {
  name: string;
  cost: number;
  duration?: number;
  description?: string;
}